pyproject.toml
[tool.poetry]
name = "sanae_bot"
version = "0.1.13"
description = "A QQBot base on nonebot2"
authors = ["AB_Solution"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
ujson = "4.0.2"
cn2an = "^0.5.9"
Pillow = "^8.1.0"
emoji = "^1.2.0"
tencentcloud-sdk-python = "^3.0.379"
lxml = "^4.6.3"
aiohttp = "^3.7.4"
DBUtils = "^2.0.1"
nb-cli = "^0.4.1"
nonebot-adapter-gocq = "^2.0.0-alpha.13"
toml = "^0.10.2"
nonebot-plugin-apscheduler = "^0.1.2"
aiomysql = "^0.0.21"
aiofiles = "^0.7.0"
opencv-python = "^4.5.3"
imageio = "^2.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
